requirements.txt
Flask
flask_login
flask_sqlalchemy
flask_wtf
flask_migrate
sqlalchemy
marshmallow
passlib
password_strength
apscheduler
email_validator
requests
opencv-python
schedule
Pillow
icmplib